{{ cookiecutter.project_slug }}/backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2
attrs==23.1.0 ; python_version >= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
cryptography==41.0.2 ; python_version >= '3.7'
defusedxml==0.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-spectacular==0.26.3
gunicorn==20.1.0
idna==3.4 ; python_version >= '3.5'
inflection==0.5.1 ; python_version >= '3.5'
jsonschema==4.18.3 ; python_version >= '3.8'
jsonschema-specifications==2023.6.1 ; python_version >= '3.8'
markdown==3.4.3
oauthlib==3.2.2 ; python_version >= '3.6'
pycparser==2.21
pyjwt==2.7.0 ; python_version >= '3.7'
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
pyyaml==6.0 ; python_version >= '3.6'
referencing==0.29.1 ; python_version >= '3.8'
requests==2.31.0 ; python_version >= '3.7'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rpds-py==0.8.10 ; python_version >= '3.8'
setuptools==68.0.0 ; python_version >= '3.7'
social-auth-app-django==5.2.0 ; python_version >= '3.7'
social-auth-core==4.4.2 ; python_version >= '3.6'
sqlparse==0.4.4
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==2.0.3 ; python_version >= '3.7'
